+++ inst/examples/ex-plot.R
@@ -32,5 +32,3 @@ viz_individuals(X, highlight = iris$Species, pch = 21:23,
 viz_variables(X, highlight = c("Petal", "Petal", "Sepal", "Sepal"),
               col = c("#EE7733", "#0077BB"),
               lty = c(1, 3))
-
-